Feat: Added avatar, signout, favorites
The log in button turns into an avatar after you log in and there's an option to sign out if you click on it. The functionality for this was also added.
When logged in, you can favorite movies by clicking on the icon and a popup will show to tell you whether you added or removed it from favorites. If not logged in, the icon and also the "My favorites" will disappear from the navbar.